Trait cerebral::neural::Neural

pub trait Neural {
    // Required methods
    fn get_lateral(&self) -> &Array2<f64>;
    fn get_lateral_mut(&mut self) -> &mut Array2<f64>;
    fn set_lateral(&mut self, lateral: Array2<f64>);
    fn get_patterns(&self) -> &Array2<f64>;
    fn get_patterns_mut(&mut self) -> &mut Array2<f64>;
    fn set_patterns(&mut self, patterns: Array2<f64>);
}

Provides access to the neurons of a neural network. The data is separated in

lateral connections: Depending on the network type, can be absolute coordinates (SOM) or connection distances (Growing neural gases) for instance.

tuning patterns: List of patterns of the feature space each individual neural is tuned to Provides read-only, modifying and consuming access.

Required Methods

fn get_lateral(&self) -> &Array2<f64>

fn get_lateral_mut(&mut self) -> &mut Array2<f64>

fn set_lateral(&mut self, lateral: Array2<f64>)

fn get_patterns(&self) -> &Array2<f64>

fn get_patterns_mut(&mut self) -> &mut Array2<f64>

fn set_patterns(&mut self, patterns: Array2<f64>)